Abstract test client and app to top level conftest.py
