+++ app/tests/conftest.py
@@ -1,0 +1,15 @@
+import pytest
+from app import create_app
+
+
+@pytest.fixture
+def app():
+    app = create_app()
+    app.config["TESTING"] = True
+    app.config["WTF_CSRF_ENABLED"] = False
+    yield app
+
+
+@pytest.fixture
+def client(app):
+    return app.test_client()
