+++ app/tests/test_search.py
@@ -1,23 +1,7 @@
 from unittest.mock import patch
 
-import pytest
 from bs4 import BeautifulSoup
 from flask.testing import FlaskClient
-
-from app import create_app
-
-
-@pytest.fixture
-def app():
-    app = create_app()
-    app.config["TESTING"] = True
-    app.config["WTF_CSRF_ENABLED"] = False
-    yield app
-
-
-@pytest.fixture
-def client(app):
-    return app.test_client()
 
 
 def test_poc_search_get(client: FlaskClient):
